requirements.txt
absl-py==2.3.1
accelerate==1.12.0
altair==6.0.0
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
astunparse==1.6.3
async-generator==1.10
async-lru==2.1.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.5
camelcase==0.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
comtypes==1.4.15
contourpy==1.3.3
cycler==0.12.1
ddgs==9.11.4
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.36.0
Django==5.2.10
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.3
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
fsspec==2026.1.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kagglehub==0.4.1
kagglesdk==0.1.15
keras==3.13.1
kiwisolver==1.4.9
langchain==1.2.13
langchain-core==1.2.20
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lark==1.3.1
libclang==18.1.1
lxml==6.0.2
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
ml_dtypes==0.5.4
mpmath==1.3.0
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.15.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
notebook==7.5.3
notebook_shim==0.2.4
numpy==2.4.1
ollama==0.6.1
opencv-python==4.13.0.90
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
ormsgpack==1.12.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pefile==2024.8.26
pillow==12.1.0
platformdirs==4.5.1
primp==1.1.3
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==23.0.0
pybind11==3.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
PythonToolsKit==1.2.6
pyttsx3==2.99
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.1
rpds-py==0.30.0
safetensors==0.7.0
scapy==2.7.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
selenium==4.40.0
Send2Trash==2.1.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlparse==0.5.5
stack-data==0.6.3
streamlit==1.53.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
terminado==0.18.1
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.5.1
tk==0.1.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.1
Tracert==0.0.1
traitlets==5.14.3
transformers==5.0.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer-slim==0.21.1
types-certifi==2021.10.8.3
types-python-dateutil==2.9.0.20260124
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uuid_utils==0.14.1
uv==0.11.6
vaderSentiment==3.3.2
watchdog==6.0.0
wcwidth==0.5.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.5
widgetsnbextension==4.0.15
word2number==1.1
wrapt==2.0.1
wsproto==1.3.2
xgboost==3.1.3
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
